remove unnecessary cast.
git-svn-id: https://geobeagle.googlecode.com/svn@1067 eab1e41a-c3b7-11dd-b16b-519e865567fa

+++ branches/andpet/GeoBeagle/src/com/google/code/geobeagle/activity/cachelist/presenter/GeocacheSummaryRowInflater.java
@@ -87,7 +87,7 @@ public class GeocacheSummaryRowInflater implements HasDistanceFormatter {
 
             final CharSequence formattedDistance = distanceFormatter
                     .formatDistance((float)distance);
-            float bearing = (float)distanceAndBearing.getBearing();
+            float bearing = distanceAndBearing.getBearing();
             final String formattedBearing = relativeBearingFormatter.formatBearing(bearing,
                     azimuth);
             return formattedDistance + " " + formattedBearing;
